+++ R/zonal_statistics.R
@@ -24,7 +24,7 @@
 #' You will most-likely want to use the spatial view layers directly e.g. with the sf package.  
 #' 
 #' 
-#' Available summary statistics currently include "min", "max", "mean", "median", "count", "sum", and "prod". 
+#' Available summary statistics currently include "min", "max", "mean", "median", "count", "sum", "prod", "var", and "sd". 
 #' 
 #' 
 #' @note Currently, the spatial reference systems of the data cube and the features must be identical.